Changes from PR review
