+++ src/app/static/src/js/utils/types.ts
@@ -2,3 +2,22 @@ export interface Parameter {
     name: string,
     value: string
 }
+
+export interface Error {
+    response?: {
+        data?: {
+            errors?: any[]
+        }
+    }
+}
+
+export interface ReportDependency {
+    id: string,
+    name: string,
+    dependencies: ReportDependency[]
+}
+
+export interface ReportDependencies {
+    direction: "upstream" | "downstream",
+    dependency_tree: ReportDependency
+}
